+++ src/app/+products/containers/product/product.component.ts
@@ -1,10 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Observable } from 'rxjs';
-
-import { ProductFacade } from 'facades/product.facade';
-import { Product } from 'models/product.model';
 import { filter, first, map, switchMap, tap } from 'rxjs/operators';
+
+import { ProductFacade } from '../../../facades';
+import { Product } from '../../../models';
 
 @Component({
   selector: 'app-product',