+++ src/app/facades/index.ts
@@ -1,0 +1,3 @@
+export * from './customer.facade';
+export * from './order.facade';
+export * from './product.facade';