+++ src/app/state/account.state.ts
@@ -1,8 +1,7 @@
 import { buildState, IEntityState } from '@briebug/ngrx-auto-entity';
-import { Account } from 'models/account.model';
+import { Account } from '../models';
 
 export const { initialState, facade: AccountFacadeBase } = buildState(Account);
-
 export function accountReducer(state = initialState): IEntityState<Account> {
   return state;
 }