+++ src/test.module.ts
@@ -3,10 +3,10 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
-import { CoreModule } from 'core/core.module';
-import { StateModule } from 'state/state.module';
 import { AppRoutingModule } from './app/app-routing.module';
+import { CoreModule } from './app/core/core.module';
 import { MaterialModule } from './app/material.module';
+import { StateModule } from './app/state/state.module';
 
 @NgModule({
   imports: [